feat: add more specific types for argTypes parameter
