+++ src/index.ts
@@ -9,3 +9,4 @@
 
 export * from './lib';
 export { withDesign } from 'storybook-addon-designs';
+export { ControlTypes, LayoutOptions } from './types/storyBook';
